Fixes for notice, payslip and extra

+++ Payhum/src/com/openhr/taxengine/impl/BaseTC.java
@@ -24,7 +24,7 @@ public class BaseTC implements TaxCalculator{
 			for(TaxRates.Slabs slab : slabs) {
 				Double toValue = slab.getTo();
 				Double fromValue = slab.getFrom() - 1;
-				Float ratePercentage = slab.getRatePercentage();
+				Double ratePercentage = slab.getRatePercentage();
 				
 				if(slab.isLast()) {
 					if(amountYetToBeTaxed == 0)
